+++ SDeMo/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## `v1.8.1`
 
+- **added** a `ceterisparibus` explanation function for CP-plots
 - **added** explanatory plots for trained models
 
 ## `v1.8.0`